Restore original float precision to TraceViewer data_curves
See #29. Should avoid unnecessary performance reduction. Also explicitly requires float64 for time vector.

+++ ephyviewer/traceviewer.py
@@ -236,8 +236,8 @@ class DataGrabber(QT.QObject):
         
         #~ print('sigs_chunk.shape', sigs_chunk.shape)
         data_curves = sigs_chunk[:, visibles].T.copy()
-        if data_curves.dtype!='float':
-            data_curves = data_curves.astype('float')
+        if data_curves.dtype!='float32':
+            data_curves = data_curves.astype('float32')
         
         if ds_ratio>1:
             
@@ -273,7 +273,7 @@ class DataGrabber(QT.QObject):
         
         #~ print(ds_ratio)
         t_start2 = self.source.index_to_time(i_start)
-        times_curves = np.arange(data_curves.shape[1], dtype='float')
+        times_curves = np.arange(data_curves.shape[1], dtype='float64') # ensure high temporal precision (see issue #28)
         times_curves /= self.source.sample_rate/ds_ratio
         if ds_ratio>1 and decimation_method == 'min_max':
             times_curves /=2
